+++ youtube-client/src/app/app.module.ts
@@ -6,13 +6,13 @@ import { MatFormFieldModule } from '@angular/material/form-field';
 import { MatInputModule } from '@angular/material/input';
 import { MatButtonModule } from '@angular/material/button';
 import AppComponent from './app.component';
-import HeaderComponent from './header/header/header.component';
-import SearchSettingsComponent from './header/search-settings/search-settings.component';
-import SearchResultsComponent from './search/search-results/search-results.component';
-import SearchItemComponent from './search/search-item/search-item.component';
-import RegistrationComponent from './main/registration/registration.component';
-import CreateCardComponent from './main/create-card/create-card.component';
-import MarkDirective from './directives/mark.directive';
+import HeaderComponent from './core/pages/header/header.component';
+import SearchSettingsComponent from './core/components/search-settings/search-settings.component';
+import SearchResultsComponent from './youtube/components/search-results/search-results.component';
+import SearchItemComponent from './youtube/components/search-item/search-item.component';
+import RegistrationComponent from './auth/components/registration/registration.component';
+import CreateCardComponent from './youtube/components/create-card/create-card.component';
+import MarkDirective from './shared/directives/mark.directive';
 
 @NgModule({
   declarations: [
