+++ youtube-client/src/app/core/components/search-settings/search-settings.component.scss
@@ -1,4 +1,4 @@
-@import '../../../styles/palette.scss';
+@import 'src/styles/palette.scss';
 
 .settings {
   display: flex;
