+++ youtube-client/src/app/core/pages/header/header.component.scss
@@ -1,4 +1,4 @@
-@import '../../../styles/_buttons';
+@import 'src/styles/_buttons';
 
 .header {
   display: flex;
