+++ youtube-client/src/app/shared/directives/mark.directive.ts
