+++ youtube-client/src/app/youtube/components/search-item/search-item.component.ts
