+++ youtube-client/src/app/youtube/components/search-results/search-results.component.ts
@@ -1,5 +1,5 @@
 import { Component, Input } from '@angular/core';
-import ISearchItem from 'src/app/models/search-item.model';
+import ISearchItem from 'src/app/youtube/models/search-item.model';
 
 @Component({
   selector: 'app-search-results',
